Move accounting utils to utils folder
